backend/requirements.txt
Django>=4.0
tensorflow>=2.12
Pillow>=9.0
numpy>=1.20
django-cors-headers>=3.0
onnxruntime>=1.15
requests